# perf: don't prettify json when not necessary