--- a/Cargo.toml
+++ b/Cargo.toml
@@ -56,6 +56,7 @@ tempfile = { version = "3.8.0", optional = true }
 fs_extra = { version = "1.3.0", optional = true }
 rand = { version = "0.8", optional = true }
 futures-util = { version = "0.3.28", optional = true }
+memmap2 = "0.9.0"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 home = "0.5.5"