Closes #23

Building foundry testdata:

pretty
3,55s user 1,54s system 300% cpu 1,698 total

no pretty
2,86s user 1,78s system 275% cpu 1,682 total